WSIO-708 revert BlogArticleHeader interface extraction due to PoC fail
