+++ applications/blogs/backend/src/main/java/pl/ds/kyanite/blogs/components/models/FeatureBlogArticleComponent.java
@@ -32,7 +32,6 @@ import org.apache.sling.models.annotations.injectorspecific.SlingObject;
 import org.apache.sling.models.annotations.injectorspecific.ValueMapValue;
 import org.apache.sling.models.factory.ModelFactory;
 import pl.ds.kyanite.blogs.components.exceptions.AuthorInfoResolvingException;
-import pl.ds.kyanite.blogs.components.models.impl.BlogArticleHeaderModelImpl;
 import pl.ds.kyanite.blogs.components.services.AuthorInfoResolverService;
 import pl.ds.kyanite.blogs.components.services.BlogArticleService;
 import pl.ds.kyanite.common.components.utils.LinkUtil;
@@ -61,7 +60,7 @@ public class FeatureBlogArticleComponent {
   private boolean isReversed;
 
   @Getter
-  private BlogArticleHeaderModelImpl blogArticleHeader;
+  private BlogArticleHeaderModel blogArticleHeader;
 
   @Getter
   private String blogArticleHeaderLink;
@@ -95,7 +94,7 @@ public class FeatureBlogArticleComponent {
   private void init() {
     final Resource blogPage = this.getArticlePage();
     if (Objects.nonNull(blogPage)) {
-      this.blogArticleHeader = modelFactory.createModel(blogPage, BlogArticleHeaderModelImpl.class);
+      this.blogArticleHeader = modelFactory.createModel(blogPage, BlogArticleHeaderModel.class);
       this.blogArticleHeaderLink = LinkUtil.handleLink(
           StringUtils.substringBefore(blogPage.getPath(), JCR_CONTENT), resourceResolver);
       this.resolveAuthorInfo(blogPage);
