--- applications/blogs/backend/src/main/java/pl/ds/kyanite/blogs/components/models/impl/BlogArticleHeaderComponentImpl.java
@@ -1,96 +1,0 @@
-/*
- * Copyright (C) 2024 Dynamic Solutions
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package pl.ds.kyanite.blogs.components.models.impl;
-
-import java.util.Objects;
-import javax.annotation.PostConstruct;
-import javax.inject.Inject;
-import lombok.Getter;
-import org.apache.sling.api.resource.Resource;
-import org.apache.sling.api.resource.ResourceResolver;
-import org.apache.sling.models.annotations.DefaultInjectionStrategy;
-import org.apache.sling.models.annotations.Model;
-import org.apache.sling.models.annotations.injectorspecific.OSGiService;
-import org.apache.sling.models.annotations.injectorspecific.SlingObject;
-import org.apache.sling.models.factory.ModelFactory;
-import pl.ds.kyanite.blogs.components.exceptions.AuthorInfoResolvingException;
-import pl.ds.kyanite.blogs.components.models.AuthorInfoModel;
-import pl.ds.kyanite.blogs.components.models.BlogArticleHeaderComponent;
-import pl.ds.kyanite.blogs.components.services.AuthorInfoResolverService;
-import pl.ds.kyanite.blogs.components.utils.ResourceUtil;
-import pl.ds.kyanite.common.components.models.layouts.TemplateHandler;
-
-@Model(
-    adaptables = { Resource.class },
-    adapters = { BlogArticleHeaderComponent.class },
-    defaultInjectionStrategy = DefaultInjectionStrategy.OPTIONAL
-)
-public class BlogArticleHeaderComponentImpl implements BlogArticleHeaderComponent {
-
-  private final Resource resource;
-
-  private final ResourceResolver resourceResolver;
-
-  private final ModelFactory modelFactory;
-
-  @Getter
-  private BlogArticleHeaderModelImpl blogArticleHeaderModel;
-
-  @Getter
-  private AuthorInfoModel authorInfoModel;
-
-  @Getter
-  private String authorInfoErrMessage;
-
-  private final AuthorInfoResolverService authorInfoResolver;
-
-  @Inject
-  public BlogArticleHeaderComponentImpl(
-      @SlingObject Resource resource,
-      @SlingObject ResourceResolver resourceResolver,
-      @OSGiService ModelFactory modelFactory,
-      @OSGiService AuthorInfoResolverService authorInfoResolver
-  ) {
-    this.resource           = resource;
-    this.resourceResolver   = resourceResolver;
-    this.modelFactory       = modelFactory;
-    this.authorInfoResolver = authorInfoResolver;
-  }
-
-  @PostConstruct
-  private void init() {
-    resolveHeaderModel();
-    resolveAuthorInfo();
-  }
-
-  private void resolveHeaderModel() {
-    final Resource currentPage = ResourceUtil.getParentPageResource(resource, resourceResolver);
-    if (Objects.nonNull(currentPage)) {
-      final Resource pageContent = ResourceUtil.getContentNode(currentPage);
-      blogArticleHeaderModel = modelFactory.createModel(
-          pageContent, BlogArticleHeaderModelImpl.class);
-    }
-  }
-
-  private void resolveAuthorInfo() {
-    try {
-      authorInfoModel = authorInfoResolver.retrieveAuthorInfo(resource, resourceResolver);
-    } catch (AuthorInfoResolvingException e) {
-      authorInfoErrMessage = e.getMessage();
-    }
-  }
-}
